[Web] WebCameraCapture 分割
